backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ng12-rag-backend"
version = "0.1.0"
description = "FastAPI backend for NG12 RAG assessment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "chromadb>=1.0.8",
  "fastapi>=0.115.0",
  "google-cloud-aiplatform>=1.67.0",
  "pypdf>=5.1.0",
  "PyYAML>=6.0.2",
  "uvicorn>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
